pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narra-tron"
version = "0.1.0"
description = "Software stack for Narra-Tron OCR + TTS + STT pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.1.0",
  "python-multipart>=0.0.20",
  "jinja2>=3.1.0",
]

[project.optional-dependencies]
ml = [
  "paddleocr>=2.8.0",
  "faster-whisper>=1.1.0",
  "paddlepaddle>=3.3.0",
  "Pillow>=10.0.0",
]
test = [
  "pytest>=8.3.0",
]

[project.scripts]
narra-tron = "narratron.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
